@@ -13,7 +13,7 @@
     }
   ],
   "require": {
-    "php": ">=7.2",
+    "php": ">=7.4",
     "ext-pdo": "*",
     "ext-json": "*",
     "ext-openssl": "*",
