@@ -125,7 +125,7 @@ abstract class Model implements \JsonSerializable, \ArrayAccess
 		{
 			return $this->$method();
 		}
-		elseif(isset($this->_data[$name]))
+		elseif(isset($this->_data[$name]) || array_key_exists($name, $this->_data))
 		{
 			return $this->_data[$name];
 		}
